refactor: extract runtime tracing middleware
